custom bg quata control for web version

+++ app/assets/js-app.js
@@ -77,6 +77,26 @@
 	_g.getStatRange = function( range ){
 		return _g.statRange[parseInt(range)]
 	};
+	
+	_g.getSize = function( bytes, target ){
+		target = target.toUpperCase()
+		
+		if( target[target.length-1] == 'B' )
+			target = target.substr(0, target.length-1)
+		
+		function _r(r){
+			return Math.floor( r * 100 ) / 100
+		}
+
+		bytes = bytes / 1024;
+		if( target == 'K' ) return _r(bytes) + 'KB';
+		bytes = bytes / 1024;
+		if( target == 'M' ) return _r(bytes) + 'MB';
+		bytes = bytes / 1024;
+		if( target == 'G' ) return _r(bytes) + 'GB';
+		bytes = bytes / 1024;
+		if( target == 'T' ) return _r(bytes) + 'TB';
+	};
 
 
 
@@ -7130,7 +7150,6 @@ BgImg.countCustom = 0;
 							return
 						}else{
 							mime = mime.split('/')
-							console.log(mime)
 							if( mime[0].toLowerCase() != 'image' ){
 								_g.badgeError('请选择图片文件')
 								_done()
@@ -7176,7 +7195,8 @@ BgImg.countCustom = 0;
 						})
 						.catch(function(err){
 							_g.error(err, '自定义背景图')
-							o.delete()
+							if( o )
+								o.delete()
 						})
 						.done(_done)
 					}
@@ -7316,7 +7336,15 @@ BgImg.countCustom = 0;
 				)
 				.append( BgImg.controlsEls.listCustom =
 					$('<dl/>',{
-						'html':	'<dt>自定义</dt>'
+						'html':	`<dt>自定义</dt>`
+					})
+					.prepend(function(){
+						if( BgImg.quota )
+							return $('<small/>')
+								.append( BgImg.controlsEls.listCustomQuotaUsed =
+									$(`<span>${_g.getSize(BgImg.quotaUsed, 'm')}</span>`)
+								)
+								.append(` / <span>${_g.getSize(BgImg.quota, 'm')}</span>`)
 					})
 					.append( BgImg.controlsEls.listCustomAdd =
 						$('<dd/>',{
